fix: Sin: static to avoid recusive resolve
There is a bug in this unmaintained plugin where sometimes, it will
check auth while resolving itself during startup, and not by the system
user.
This makes it recurse back to reloading and resolving itself, which
causes it to repeat until stack overflow.
Sinful static boolean to break that recursion one time